requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.53.0
anyio==4.9.0
attrs==25.3.0
autogen-agentchat==0.6.1
autogen-core==0.6.1
autogen-ext==0.6.1
av==14.4.0
azure-ai-agents==1.0.1
azure-ai-projects==1.0.0b11
azure-core==1.34.0
azure-identity==1.23.0
azure-storage-blob==12.25.1
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
build==1.3.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chromadb==1.0.20
click==8.2.1
cloudevents==1.12.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
cryptography==45.0.4
dataclasses-json==0.6.7
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
durationpy==0.10
ecdsa==0.19.1
fastapi==0.115.12
ffmpy==0.6.0
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
google-auth==2.40.3
google-crc32c==1.7.1
googleapis-common-protos==1.70.0
gradio==5.33.1
gradio_client==1.10.3
greenlet==3.2.3
griffe==1.7.3
groovy==0.1.2
grpcio==1.70.0
h11==0.16.0
hf-xet==1.1.3
html2text==2025.4.15
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.4
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
importlib_resources==6.5.2
ipywidgets==8.1.7
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
jupyterlab_widgets==3.0.15
kubernetes==33.1.0
langchain==0.3.25
langchain-anthropic==0.3.15
langchain-community==0.3.24
langchain-core==0.3.64
langchain-experimental==0.3.4
langchain-openai==0.3.21
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-checkpoint-sqlite==2.0.10
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
lazy-object-proxy==1.11.0
lxml==5.4.0
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.9.3
mcp-server-fetch==2025.1.17
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.7.0
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.42.0
numpy==2.3.0
oauthlib==3.3.1
ollama==0.5.1
onnxruntime==1.22.1
openai==1.85.0
openai-agents==0.0.17
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.10.18
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==2.3.0
parse==1.20.2
pathable==0.4.4
pillow==11.2.1
playwright==1.52.0
plotly==6.1.2
polygon-api-client==1.14.6
posthog==5.4.0
prance==25.4.8.0
propcache==0.3.2
Protego==0.4.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybars4==0.9.13
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.12.0
PyMeta3==0.5.1
pyOpenSSL==25.1.0
pypdf==5.6.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dotenv==1.1.0
python-http-client==3.3.7
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
readabilipy==0.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
ruff==0.11.13
safehttpx==0.1.6
scikit-learn==1.7.1
scipy==1.15.3
semantic-kernel==1.32.2
semantic-version==2.10.0
sendgrid==6.12.3
shellingham==1.5.4
smithery==0.1.0
sniffio==1.3.1
soupsieve==2.7
speedtest-cli==2.1.3
SQLAlchemy==2.0.41
sqlite-vec==0.1.6
sse-starlette==2.3.6
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.4
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.0
types-requests==2.32.0.20250602
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.1
widgetsnbextension==4.0.14
wikipedia==1.4.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
ipykernel
docling
lancedb
streamlit
flask
ddgs
matplotlib.pyplot 
seaborn

